+++ crutch-tv/crutch-tv/CrutchTVApp.swift
@@ -1,0 +1,12 @@
+import SwiftUI
+
+@main
+struct CrutchTVApp: App {
+    var body: some Scene {
+        WindowGroup {
+            NavigationStack {
+                SongsListView()
+            }
+        }
+    }
+}